Contact-Us
The Contact Web page should bring up a form which allows users to   send
an email to your UTS email address. The form should have a field for the
subject of the email, the user's email address, the user's first name,
the   user's last name, a textarea for the contents of the email and a
submit   button. The subject fields, user's email address and textarea
are all   compulsory fields. When the submit button is pressed a
confirmation email   should be sent to the email address the user
supplied.